+++ packages/ui/src/index.ts
@@ -38,6 +38,7 @@ export * from "./components/multi-select";
 export * from "./components/radio";
 export * from "./components/resizable-screen";
 export * from "./components/select";
+export type { SelectMenuOption } from "./components/select-menu";
 export * from "./components/sidebar";
 export * from "./components/spinner";
 export * from "./components/switch";